+++ finality-provider/cmd/fpd/common/keys_cmd.go
@@ -1,10 +1,16 @@
-package daemon
+package common
 
 import (
 	"github.com/babylonlabs-io/finality-provider/util"
 	"github.com/cosmos/cosmos-sdk/client/keys"
 	"github.com/spf13/cobra"
 )
+
+// AddKeysCommands adds all the keys-related commands to the provided command.
+// The keys commands are generic to {Babylon, Cosmos BSN, rollup BSN} finality providers
+func AddKeysCommands(cmd *cobra.Command) {
+	cmd.AddCommand(CommandKeys())
+}
 
 // CommandKeys returns the keys group command and updates the add command to do a
 // post run action to update the config if exists.