+++ finality-provider/cmd/fpd/daemon/start.go
@@ -6,31 +6,35 @@ import (
 	"path/filepath"
 
 	"github.com/babylonlabs-io/babylon/v3/types"
-	"github.com/btcsuite/btcwallet/walletdb"
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cobra"
-	"go.uber.org/zap"
-
-	fpcmd "github.com/babylonlabs-io/finality-provider/finality-provider/cmd"
+	clientctx "github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/clientctx"
+	commoncmd "github.com/babylonlabs-io/finality-provider/finality-provider/cmd/fpd/common"
 	fpcfg "github.com/babylonlabs-io/finality-provider/finality-provider/config"
 	"github.com/babylonlabs-io/finality-provider/finality-provider/service"
 	"github.com/babylonlabs-io/finality-provider/log"
 	"github.com/babylonlabs-io/finality-provider/util"
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/spf13/cobra"
 )
 
 // CommandStart returns the start command of fpd daemon.
-func CommandStart() *cobra.Command {
+func CommandStart(binaryName string) *cobra.Command {
+	cmd := CommandStartTemplate(binaryName)
+	cmd.RunE = clientctx.RunEWithClientCtx(runStartCmd)
+
+	return cmd
+}
+
+func CommandStartTemplate(binaryName string) *cobra.Command {
 	var cmd = &cobra.Command{
 		Use:     "start",
 		Short:   "Start the finality-provider app daemon.",
 		Long:    `Start the finality-provider app. Note that eotsd should be started beforehand`,
-		Example: `fpd start --home /home/user/.fpd`,
+		Example: fmt.Sprintf(`%s start --home /home/user/.fpd`, binaryName),
 		Args:    cobra.NoArgs,
-		RunE:    fpcmd.RunEWithClientCtx(runStartCmd),
 	}
-	cmd.Flags().String(fpEotsPkFlag, "", "The EOTS public key of the finality-provider to start")
-	cmd.Flags().String(rpcListenerFlag, "", "The address that the RPC server listens to")
+	cmd.Flags().String(commoncmd.FpEotsPkFlag, "", "The EOTS public key of the finality-provider to start")
+	cmd.Flags().String(commoncmd.RPCListenerFlag, "", "The address that the RPC server listens to")
 	cmd.Flags().String(flags.FlagHome, fpcfg.DefaultFpdDir, "The application home directory")
 
 	return cmd
@@ -42,21 +46,21 @@ func runStartCmd(ctx client.Context, cmd *cobra.Command, _ []string) error {
 		return err
 	}
 	homePath = util.CleanAndExpandPath(homePath)
-	flags := cmd.Flags()
-
-	fpStr, err := flags.GetString(fpEotsPkFlag)
-	if err != nil {
-		return fmt.Errorf("failed to read flag %s: %w", fpEotsPkFlag, err)
-	}
-
-	rpcListener, err := flags.GetString(rpcListenerFlag)
-	if err != nil {
-		return fmt.Errorf("failed to read flag %s: %w", rpcListenerFlag, err)
-	}
-
 	cfg, err := fpcfg.LoadConfig(homePath)
 	if err != nil {
 		return fmt.Errorf("failed to load configuration: %w", err)
+	}
+
+	flags := cmd.Flags()
+
+	fpStr, err := flags.GetString(commoncmd.FpEotsPkFlag)
+	if err != nil {
+		return fmt.Errorf("failed to read flag %s: %w", commoncmd.FpEotsPkFlag, err)
+	}
+
+	rpcListener, err := flags.GetString(commoncmd.RPCListenerFlag)
+	if err != nil {
+		return fmt.Errorf("failed to read flag %s: %w", commoncmd.RPCListenerFlag, err)
 	}
 
 	if cfg.BabylonConfig.KeyringBackend != "test" {
@@ -81,12 +85,12 @@ func runStartCmd(ctx client.Context, cmd *cobra.Command, _ []string) error {
 		return fmt.Errorf("failed to create db backend: %w", err)
 	}
 
-	fpApp, err := loadApp(logger, cfg, dbBackend)
+	fpApp, err := service.NewFinalityProviderAppFromConfig(cfg, dbBackend, logger)
 	if err != nil {
-		return fmt.Errorf("failed to load app: %w", err)
+		return fmt.Errorf("failed to create finality-provider app: %w", err)
 	}
 
-	if err := startApp(fpApp, fpStr); err != nil {
+	if err := StartApp(fpApp, fpStr); err != nil {
 		return fmt.Errorf("failed to start app: %w", err)
 	}
 
@@ -95,22 +99,8 @@ func runStartCmd(ctx client.Context, cmd *cobra.Command, _ []string) error {
 	return fpServer.RunUntilShutdown(cmd.Context())
 }
 
-// loadApp initializes a finality provider app based on config and flags set.
-func loadApp(
-	logger *zap.Logger,
-	cfg *fpcfg.Config,
-	dbBackend walletdb.DB,
-) (*service.FinalityProviderApp, error) {
-	fpApp, err := service.NewFinalityProviderAppFromConfig(cfg, dbBackend, logger)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create finality-provider app: %w", err)
-	}
-
-	return fpApp, nil
-}
-
-// startApp starts the app and the handle of finality providers if needed based on flags.
-func startApp(
+// StartApp starts the app and the handle of finality providers if needed based on flags.
+func StartApp(
 	fpApp *service.FinalityProviderApp,
 	fpPkStr string,
 ) error {