+++ CHANGELOG.md
@@ -43,14 +43,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
-* [#517](https://github.com/babylonlabs-io/finality-provider/pull/518) feat: rollup BSN finality provider setup
 * [#493](https://github.com/babylonlabs-io/finality-provider/pull/493) chore: key migration for test keyring
 * [#501](https://github.com/babylonlabs-io/finality-provider/pull/501) chore: signing context
 * [#500](https://github.com/babylonlabs-io/finality-provider/pull/500) feat: fpd abstract poller implementation
 * [502](https://github.com/babylonlabs-io/finality-provider/pull/502) chore: bump babylon to v3
 * [512](https://github.com/babylonlabs-io/finality-provider/pull/512) chore: consumer interface refactor
 * [515](https://github.com/babylonlabs-io/finality-provider/pull/515) chore: abstract randomness committer
+* [#522](https://github.com/babylonlabs-io/finality-provider/pull/522) cli: refactor CLIs and generalise for rollup FPs
 * [516](https://github.com/babylonlabs-io/finality-provider/pull/516) chore: abstract bootstrap
+* [#517](https://github.com/babylonlabs-io/finality-provider/pull/518) feat: rollup BSN finality provider setup
+* [#526](https://github.com/babylonlabs-io/finality-provider/pull/526) rollup: remove dependency to finality gadget
+* [537](https://github.com/babylonlabs-io/finality-provider/pull/537) chore: mv helper methods from fpi
 * [520](https://github.com/babylonlabs-io/finality-provider/pull/520) chore: migration guide keyring test to file
 
 ## v1.1.0-rc.1